+++ spp_registry/tests/__init__.py
@@ -10,6 +10,7 @@ from . import test_individual_name
 from . import test_phone_number
 from . import test_reg_id
 from . import test_membership_constraints
+from . import test_membership_status_cron
 from . import test_registrant_misc
 from . import test_group_aggregation
 from . import test_res_config_settings